requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.29
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-cors-headers==3.13.0
idna==3.3
numpy==1.22.4
pandas==1.4.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
djangorestframework==3.14.0